types

+++ test/backend_test.in.cpp
@@ -309,12 +309,12 @@ TEST(${BACKEND_NAME}, batch_norm_one_output)
     auto backend = runtime::Backend::create("${BACKEND_NAME}");
 
     // Create some tensors for input/output
-    auto a = backend->create_tensor(element::f32, shape_in);
+    auto a = backend->create_tensor(element::f64, shape_in);
     copy_data(
         a,
         vector<double>{-1.97431703, -2.06521307, 0.54122217, 2.53375939, -0.22342691, 0.45340773});
 
-    auto result = backend->create_tensor(element::f32, shape_in);
+    auto result = backend->create_tensor(element::f64, shape_in);
     vector<double> expected_result{
         -0.09365749, -1.01327395, -1.04269195, 5.00118923, -0.43295258, -1.24840283};
 
@@ -348,7 +348,7 @@ TEST(${BACKEND_NAME}, batch_norm_three_outputs)
     auto backend = runtime::Backend::create("${BACKEND_NAME}");
 
     // Create some tensors for input/output
-    auto a = backend->create_tensor(element::f32, shape_in);
+    auto a = backend->create_tensor(element::f64, shape_in);
     copy_data(
         a,
         vector<double>{-1.97431703, -2.06521307, 0.54122217, 2.53375939, -0.22342691, 0.45340773});
@@ -360,13 +360,13 @@ TEST(${BACKEND_NAME}, batch_norm_three_outputs)
     backend->call(f0, {result0}, {a});
     EXPECT_TRUE(test::all_close(vector<double>{expected_result0}, read_vector<double>(result0)));
 
-    auto result1 = backend->create_tensor(element::f32, shape_mean);
+    auto result1 = backend->create_tensor(element::f64, shape_mean);
     vector<double> expected_result1{0.27972114, -1.14431989, 0.49731493};
 
     backend->call(f1, {result1}, {a});
     EXPECT_TRUE(test::all_close(vector<double>{expected_result1}, read_vector<double>(result1)));
 
-    auto result2 = backend->create_tensor(element::f32, shape_mean);
+    auto result2 = backend->create_tensor(element::f64, shape_mean);
     vector<double> expected_result2{5.08068895e+00, 8.48043919e-01, 1.92784308e-03};
 
     backend->call(f2, {result2}, {a});
